+++ Tests/demistomock/demistomock.py
@@ -555,7 +555,10 @@ def handleEntitlementForUser(incidentID, guid, email, content, taskID=""):
     return {}
 
 def demistoVersion():
-    return {}
+    return {
+        'version': '5.5.0',
+        'buildNumber': '12345'
+    }
 
 def integrationInstance():
     return ""