+++ Packs/GenericSQL/Integrations/GenericSQL/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
-
+- Fixed an issue where mysql default tables where not available for querying.
+- Added support for stack trace and the query error message.
+- Added support to Oracle connection.
+- Added support for **limit** and **skip** command arguments in context output.
 
 ## [20.3.4] - 2020-03-30
 -